+++ index.html
@@ -227,7 +227,7 @@
     </section>
   </template>
 
+  <script src="js/main.js" type="module"></script>
   <script src="js/function.js"></script>
-  <script src="js/main.js"></script>
 </body>
 </html>
